requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
chardet==4.0.0
click==8.0.3
discord-py-slash-command==3.0.1
discord.py==1.7.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
soupsieve==2.3.1
typing_extensions==4.0.1
Werkzeug==2.0.2
yarl==1.7.2